i2c2 to i2c
