+++ include/ms5611.h
@@ -33,7 +33,7 @@
 #define MS5611_H
 
 #include "system.h"
-#include "i2c2.h"
+#include "i2c.h"
 
 
 class MS5611
@@ -83,7 +83,7 @@ private:
   bool start_pres_meas();
   void convert();
 
-  i2c2::I2C* i2c_;
+  I2C* i2c_;
   uint8_t pres_buf_[3];
   uint8_t temp_buf_[3];
   int32_t pres_raw_;
@@ -101,7 +101,7 @@ private:
   callback_type_t callback_type_;
 
 public:
-  bool init(i2c2::I2C* _i2c);
+  bool init(I2C* _i2c);
   void update();
   void read(float *press, float *temp);
   bool present();
